+++ src/software-delivery-machine/blueprint/deploy/k8sDeploy.ts
@@ -1,6 +1,6 @@
 import { NoticeK8sTestDeployCompletionOnStatus } from "../../../handlers/events/delivery/deploy/k8s/NoticeK8sDeployCompletion";
 import { NoticeK8sProdDeployCompletionOnStatus } from "../../../handlers/events/delivery/deploy/k8s/NoticeK8sProdDeployCompletion";
-import { requestDeployToK8s } from "../../../handlers/events/delivery/deploy/k8s/RequestK8sDeployOnSuccessStatus";
+import { requestDeployToK8s } from "../../../handlers/events/delivery/deploy/k8s/RequestK8sDeploys";
 import {
     ProductionDeploymentGoal,
     ProductionEndpointGoal,
